move profile and whatnot to Forms folder
